feat: DefaultPostCard 컴포넌트 구현 (#21)
- 여러 형태의 PostCard가 존재할 수 있기 때문에, `/presets` 폴더로 분리 (추후 다른 형태 추가)
